Add hooks for clients, gigs and invoices
