+++ frontend/glovelly-web/src/hooks/useClientsWorkspace.ts
@@ -1,0 +1,450 @@
+import { useCallback, useDeferredValue, useMemo, useState } from 'react'
+import type { FormEvent } from 'react'
+import {
+  buildApiUrl,
+  emptyClientSettingsForm,
+  emptyForm,
+  fetchWithSession,
+  parseProblemDetails,
+} from '../appShared'
+import type { Address, Client, ClientForm, ClientSettingsForm } from '../appShared'
+
+type UseClientsWorkspaceOptions = {
+  isApiConnected: boolean
+  onSessionExpired: (message: string) => void
+  setIsLoading: (isLoading: boolean) => void
+  setStatus: (status: string) => void
+}
+
+function shouldCloseAfterSave(event: FormEvent<HTMLFormElement>) {
+  const submitter = (event.nativeEvent as SubmitEvent).submitter as
+    | HTMLButtonElement
+    | null
+
+  return submitter?.dataset.closeAfterSave !== 'false'
+}
+
+function parseOptionalDecimal(value: string) {
+  const trimmed = value.trim()
+  if (!trimmed) {
+    return null
+  }
+
+  const parsed = Number(trimmed)
+  return Number.isFinite(parsed) ? parsed : Number.NaN
+}
+
+export function useClientsWorkspace({
+  isApiConnected,
+  onSessionExpired,
+  setIsLoading,
+  setStatus,
+}: UseClientsWorkspaceOptions) {
+  const [clients, setClients] = useState<Client[]>([])
+  const [selectedClientId, setSelectedClientId] = useState<string>('')
+  const [searchQuery, setSearchQuery] = useState('')
+  const [isClientEditorOpen, setIsClientEditorOpen] = useState(false)
+  const [mode, setMode] = useState<'create' | 'edit'>('create')
+  const [form, setForm] = useState<ClientForm>(emptyForm)
+  const [isClientSettingsOpen, setIsClientSettingsOpen] = useState(false)
+  const [clientSettingsForm, setClientSettingsForm] =
+    useState<ClientSettingsForm>(emptyClientSettingsForm)
+  const [clientSettingsStatus, setClientSettingsStatus] =
+    useState('Client-specific rates override your personal defaults when set.')
+  const [isClientSettingsSaving, setIsClientSettingsSaving] = useState(false)
+  const deferredSearchQuery = useDeferredValue(searchQuery)
+
+  const clientsById = useMemo(
+    () => new Map(clients.map((client) => [client.id, client])),
+    [clients]
+  )
+  const clientNamesById = useMemo(
+    () => new Map(clients.map((client) => [client.id, client.name])),
+    [clients]
+  )
+
+  const filteredClients = useMemo(() => {
+    const query = deferredSearchQuery.trim().toLowerCase()
+    if (!query) {
+      return clients
+    }
+
+    return clients.filter((client) =>
+      [
+        client.name,
+        client.email,
+        client.billingAddress.city,
+        client.billingAddress.country,
+      ]
+        .join(' ')
+        .toLowerCase()
+        .includes(query)
+    )
+  }, [clients, deferredSearchQuery])
+
+  const selectedClient = clientsById.get(selectedClientId) ?? filteredClients[0] ?? null
+
+  const applyClients = useCallback((nextClients: Client[]) => {
+    setClients(nextClients)
+    setSelectedClientId(nextClients[0]?.id ?? '')
+  }, [])
+
+  const resetClientsWorkspace = useCallback(() => {
+    setClients([])
+    setSelectedClientId('')
+    setSearchQuery('')
+    setIsClientEditorOpen(false)
+    setMode('create')
+    setForm(emptyForm())
+    setIsClientSettingsOpen(false)
+    setClientSettingsForm(emptyClientSettingsForm())
+    setClientSettingsStatus(
+      'Client-specific rates override your personal defaults when set.'
+    )
+    setIsClientSettingsSaving(false)
+  }, [])
+
+  const startCreating = () => {
+    setMode('create')
+    setForm(emptyForm())
+    setIsClientEditorOpen(true)
+  }
+
+  const startEditing = () => {
+    if (!selectedClient) {
+      return
+    }
+
+    setMode('edit')
+    setForm({
+      name: selectedClient.name,
+      email: selectedClient.email,
+      billingAddress: {
+        line1: selectedClient.billingAddress.line1 ?? '',
+        line2: selectedClient.billingAddress.line2 ?? '',
+        city: selectedClient.billingAddress.city ?? '',
+        stateOrCounty: selectedClient.billingAddress.stateOrCounty ?? '',
+        postalCode: selectedClient.billingAddress.postalCode ?? '',
+        country: selectedClient.billingAddress.country ?? '',
+      },
+    })
+    setIsClientEditorOpen(true)
+  }
+
+  const closeClientEditor = () => {
+    setIsClientEditorOpen(false)
+    setMode('create')
+    setForm(emptyForm())
+  }
+
+  const updateField = (field: keyof ClientForm, value: string | Address) => {
+    setForm((current) => ({
+      ...current,
+      [field]: value,
+    }))
+  }
+
+  const updateAddressField = (field: keyof Address, value: string) => {
+    setForm((current) => ({
+      ...current,
+      billingAddress: {
+        ...current.billingAddress,
+        [field]: value,
+      },
+    }))
+  }
+
+  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault()
+    const closeAfterSave = shouldCloseAfterSave(event)
+
+    const payload: ClientForm = {
+      name: form.name.trim(),
+      email: form.email.trim(),
+      billingAddress: {
+        line1: form.billingAddress.line1.trim(),
+        line2: form.billingAddress.line2.trim(),
+        city: form.billingAddress.city.trim(),
+        stateOrCounty: form.billingAddress.stateOrCounty.trim(),
+        postalCode: form.billingAddress.postalCode.trim(),
+        country: form.billingAddress.country.trim(),
+      },
+    }
+
+    const preservedClientSettings =
+      mode === 'edit' && selectedClient
+        ? {
+            mileageRate: selectedClient.mileageRate,
+            passengerMileageRate: selectedClient.passengerMileageRate,
+            invoiceFilenamePattern: selectedClient.invoiceFilenamePattern,
+            invoiceEmailSubjectPattern: selectedClient.invoiceEmailSubjectPattern,
+          }
+        : {
+            mileageRate: null,
+            passengerMileageRate: null,
+            invoiceFilenamePattern: null,
+            invoiceEmailSubjectPattern: null,
+          }
+
+    if (
+      !payload.name ||
+      !payload.email ||
+      !payload.billingAddress.line1 ||
+      !payload.billingAddress.city
+    ) {
+      setStatus('Name, email, address line 1 and city are required.')
+      return
+    }
+
+    setIsLoading(true)
+
+    try {
+      if (!isApiConnected) {
+        throw new Error('API unavailable.')
+      }
+
+      const isEdit = mode === 'edit' && selectedClient
+      const endpoint = isEdit
+        ? buildApiUrl(`/clients/${selectedClient.id}`)
+        : buildApiUrl('/clients')
+      const requestBody = JSON.stringify({
+        ...payload,
+        ...preservedClientSettings,
+      })
+
+      const response = await fetchWithSession(endpoint, {
+        method: isEdit ? 'PUT' : 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: requestBody,
+      })
+
+      if (response.status === 401) {
+        onSessionExpired('Your session expired. Sign in again to save changes.')
+        return
+      }
+
+      if (!response.ok) {
+        throw new Error('Save failed.')
+      }
+
+      const savedClient = (await response.json()) as Client
+
+      setClients((current) => {
+        if (isEdit) {
+          return current.map((client) =>
+            client.id === savedClient.id ? savedClient : client
+          )
+        }
+
+        return [savedClient, ...current]
+      })
+
+      setSelectedClientId(savedClient.id)
+      setMode('edit')
+      setStatus(isEdit ? 'Client updated.' : 'Client created.')
+      setIsClientEditorOpen(!closeAfterSave)
+    } catch {
+      setStatus('Unable to save right now. Please try again.')
+    } finally {
+      setIsLoading(false)
+    }
+  }
+
+  const handleDelete = async () => {
+    if (!selectedClient) {
+      return
+    }
+
+    setIsLoading(true)
+
+    try {
+      if (!isApiConnected) {
+        throw new Error('API unavailable.')
+      }
+
+      const response = await fetchWithSession(buildApiUrl(`/clients/${selectedClient.id}`), {
+        method: 'DELETE',
+      })
+
+      if (response.status === 401) {
+        onSessionExpired('Your session expired. Sign in again to delete clients.')
+        return
+      }
+
+      if (!response.ok) {
+        throw new Error('Delete failed.')
+      }
+
+      let nextSelectedClientId = ''
+
+      setClients((current) => {
+        const remaining = current.filter((client) => client.id !== selectedClient.id)
+        nextSelectedClientId = remaining[0]?.id ?? ''
+        return remaining
+      })
+
+      setSelectedClientId(nextSelectedClientId)
+      setMode('create')
+      setForm(emptyForm())
+      setStatus('Client deleted.')
+      setIsClientEditorOpen(false)
+    } catch {
+      setStatus('Unable to delete right now.')
+    } finally {
+      setIsLoading(false)
+    }
+  }
+
+  const openClientSettings = () => {
+    if (!selectedClient) {
+      return
+    }
+
+    setClientSettingsForm({
+      mileageRate:
+        selectedClient.mileageRate === null ? '' : String(selectedClient.mileageRate),
+      passengerMileageRate:
+        selectedClient.passengerMileageRate === null
+          ? ''
+          : String(selectedClient.passengerMileageRate),
+      invoiceFilenamePattern: selectedClient.invoiceFilenamePattern ?? '',
+      invoiceEmailSubjectPattern: selectedClient.invoiceEmailSubjectPattern ?? '',
+    })
+    setClientSettingsStatus(
+      'Client-specific rates override your personal defaults when set.'
+    )
+    setIsClientSettingsOpen(true)
+  }
+
+  const closeClientSettings = () => {
+    setIsClientSettingsOpen(false)
+  }
+
+  const updateClientSettingsField = (
+    field: keyof ClientSettingsForm,
+    value: string
+  ) => {
+    setClientSettingsForm((current) => ({
+      ...current,
+      [field]: value,
+    }))
+  }
+
+  const handleClientSettingsSubmit = async (
+    event: FormEvent<HTMLFormElement>
+  ) => {
+    event.preventDefault()
+
+    if (!selectedClient) {
+      return
+    }
+
+    const mileageRate = parseOptionalDecimal(clientSettingsForm.mileageRate)
+    const passengerMileageRate = parseOptionalDecimal(
+      clientSettingsForm.passengerMileageRate
+    )
+    const invoiceFilenamePattern = clientSettingsForm.invoiceFilenamePattern.trim()
+    const invoiceEmailSubjectPattern =
+      clientSettingsForm.invoiceEmailSubjectPattern.trim()
+
+    if (Number.isNaN(mileageRate) || Number.isNaN(passengerMileageRate)) {
+      setClientSettingsStatus('Rates must be valid numbers, for example 0.45.')
+      return
+    }
+
+    setIsClientSettingsSaving(true)
+
+    try {
+      const response = await fetchWithSession(
+        buildApiUrl(`/clients/${selectedClient.id}`),
+        {
+          method: 'PUT',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({
+            id: selectedClient.id,
+            name: selectedClient.name,
+            email: selectedClient.email,
+            billingAddress: selectedClient.billingAddress,
+            mileageRate,
+            passengerMileageRate,
+            invoiceFilenamePattern: invoiceFilenamePattern || null,
+            invoiceEmailSubjectPattern: invoiceEmailSubjectPattern || null,
+          }),
+        }
+      )
+
+      if (response.status === 401) {
+        setIsClientSettingsOpen(false)
+        onSessionExpired('Your session expired. Sign in again to update client settings.')
+        return
+      }
+
+      if (!response.ok) {
+        const problem = await parseProblemDetails(response)
+        const validationMessages = problem?.errors
+          ? Object.values(problem.errors).flat().join(' ')
+          : problem?.detail ?? problem?.title
+
+        throw new Error(validationMessages || 'Unable to save client settings.')
+      }
+
+      const savedClient = (await response.json()) as Client
+
+      setClients((current) =>
+        current.map((client) => (client.id === savedClient.id ? savedClient : client))
+      )
+      setClientSettingsForm({
+        mileageRate: savedClient.mileageRate === null ? '' : String(savedClient.mileageRate),
+        passengerMileageRate:
+          savedClient.passengerMileageRate === null
+            ? ''
+            : String(savedClient.passengerMileageRate),
+        invoiceFilenamePattern: savedClient.invoiceFilenamePattern ?? '',
+        invoiceEmailSubjectPattern: savedClient.invoiceEmailSubjectPattern ?? '',
+      })
+      setClientSettingsStatus('Client settings updated.')
+    } catch (error) {
+      setClientSettingsStatus(
+        error instanceof Error
+          ? error.message
+          : 'Unable to save client settings right now.'
+      )
+    } finally {
+      setIsClientSettingsSaving(false)
+    }
+  }
+
+  return {
+    applyClients,
+    clientNamesById,
+    clientSettingsForm,
+    clientSettingsStatus,
+    clients,
+    closeClientEditor,
+    closeClientSettings,
+    filteredClients,
+    form,
+    handleClientSettingsSubmit,
+    handleDelete,
+    handleSubmit,
+    isClientEditorOpen,
+    isClientSettingsOpen,
+    isClientSettingsSaving,
+    mode,
+    openClientSettings,
+    resetClientsWorkspace,
+    searchQuery,
+    selectedClient,
+    setSelectedClientId,
+    setSearchQuery,
+    startCreating,
+    startEditing,
+    updateAddressField,
+    updateClientSettingsField,
+    updateField,
+  }
+}
